Adding expand-on-click fn to the child-row-opts
This change adds a new optional parameter passed to the child-row-opts.
The new parameter is a function which is invoked on the 'on-click' event
of the child-row's expand button. The function takes a row-data and a
boolean with the status of the child row (expanded/collapsed). This
function can be used to generate side-effect based on the status and
values of a single child row.